Refactor project structure: Organize modules into core and common directories
