+++ PROJECT_STRUCTURE_OPTIMIZATION.md
@@ -1,0 +1,183 @@
+# Service Vitals 项目结构优化总结
+
+## 问题分析
+
+### 原始问题
+在重构 `main.rs` 文件后，`src/` 根目录下仍然有 9 个 `.rs` 文件，存在以下问题：
+
+1. **文件过多**：根目录文件数量过多，影响可读性
+2. **职责分散**：相关功能的文件分散在不同位置
+3. **维护困难**：难以快速定位相关功能
+4. **不符合最佳实践**：违反了模块化组织原则
+
+### 文件分布问题
+```
+src/
+├── main.rs (311B)              # 入口点
+├── lib.rs (911B)               # 库入口
+├── app.rs (4.8KB)              # 应用程序逻辑
+├── service.rs (7.5KB)          # 服务管理
+├── daemon_service.rs (3.1KB)   # 守护进程服务
+├── foreground_service.rs (2.5KB) # 前台服务
+├── error.rs (7.7KB)            # 错误处理
+├── logging.rs (26KB)           # 日志系统
+├── status.rs (8.2KB)           # 状态管理
+├── cli/                        # CLI模块
+├── config/                     # 配置模块
+├── daemon/                     # 守护进程模块
+├── health/                     # 健康检测模块
+├── notification/               # 通知模块
+└── web/                        # Web模块
+```
+
+## 优化方案
+
+### 1. 模块化重组
+
+将相关功能的文件组织到逻辑模块中：
+
+#### `src/core/` - 核心应用程序模块
+包含应用程序的核心逻辑和生命周期管理：
+- `app.rs` - 应用程序核心逻辑
+- `service.rs` - 服务管理
+- `daemon_service.rs` - 守护进程服务
+- `foreground_service.rs` - 前台服务
+- `mod.rs` - 模块导出
+
+#### `src/common/` - 通用功能模块
+包含跨模块使用的通用功能：
+- `error.rs` - 错误处理
+- `logging.rs` - 日志系统
+- `status.rs` - 状态管理
+- `mod.rs` - 模块导出
+
+### 2. 优化后的结构
+
+```
+src/
+├── main.rs                     # 入口点 (15行)
+├── lib.rs                      # 库入口 (37行)
+├── core/                       # 核心应用程序模块
+│   ├── mod.rs                  # 模块导出
+│   ├── app.rs                  # 应用程序逻辑 (153行)
+│   ├── service.rs              # 服务管理 (245行)
+│   ├── daemon_service.rs       # 守护进程服务 (99行)
+│   └── foreground_service.rs   # 前台服务 (88行)
+├── common/                     # 通用功能模块
+│   ├── mod.rs                  # 模块导出
+│   ├── error.rs                # 错误处理 (270行)
+│   ├── logging.rs              # 日志系统 (831行)
+│   └── status.rs               # 状态管理 (271行)
+├── cli/                        # CLI模块
+├── config/                     # 配置模块
+├── daemon/                     # 守护进程模块
+├── health/                     # 健康检测模块
+├── notification/               # 通知模块
+└── web/                        # Web模块
+```
+
+## 优化效果
+
+### 1. 根目录文件数量减少
+- **优化前**：9 个 `.rs` 文件
+- **优化后**：2 个 `.rs` 文件（`main.rs` 和 `lib.rs`）
+
+### 2. 逻辑分组清晰
+- **核心功能**：应用程序逻辑集中在 `core/` 模块
+- **通用功能**：跨模块功能集中在 `common/` 模块
+- **业务功能**：各业务模块保持独立
+
+### 3. 职责分离明确
+- `main.rs`：只负责程序入口
+- `lib.rs`：只负责模块导出和公共API
+- `core/`：应用程序核心逻辑
+- `common/`：通用工具和基础设施
+
+### 4. 可维护性提升
+- 相关功能文件集中管理
+- 模块边界清晰
+- 依赖关系简化
+
+## 模块职责说明
+
+### `src/core/` 模块
+负责应用程序的核心逻辑和生命周期管理：
+
+- **`app.rs`**：应用程序主函数、命令执行逻辑
+- **`service.rs`**：服务管理器、服务启动器、组件初始化
+- **`daemon_service.rs`**：守护进程模式的启动和管理
+- **`foreground_service.rs`**：前台模式的启动和信号处理
+
+### `src/common/` 模块
+提供跨模块使用的通用功能：
+
+- **`error.rs`**：统一错误类型、错误处理机制
+- **`logging.rs`**：日志系统配置和管理
+- **`status.rs`**：服务状态管理和持久化
+
+## 导入路径更新
+
+### 主要变更
+```rust
+// 优化前
+use service_vitals::app;
+use service_vitals::error::ServiceVitalsError;
+
+// 优化后
+use service_vitals::core::app;
+use service_vitals::common::error::ServiceVitalsError;
+```
+
+### 重新导出
+在 `lib.rs` 中重新导出常用类型，保持向后兼容：
+```rust
+pub use common::error::ServiceVitalsError;
+pub use common::logging::{LogConfig, LoggingSystem};
+pub use common::status::{ServiceStatus, StatusManager};
+```
+
+## 最佳实践遵循
+
+### 1. 模块化原则
+- 单一职责原则：每个模块职责明确
+- 高内聚低耦合：相关功能集中，模块间依赖清晰
+- 分层架构：核心、通用、业务功能分层
+
+### 2. Rust 项目规范
+- 目录结构清晰：`src/` 下只保留入口文件
+- 模块组织合理：相关功能组织到同一模块
+- 命名规范：使用 `snake_case` 命名目录和文件
+
+### 3. 可扩展性
+- 模块化设计便于添加新功能
+- 清晰的模块边界便于重构
+- 统一的导入路径便于维护
+
+## 后续建议
+
+### 1. 进一步优化
+- 考虑将 `logging.rs`（831行）进一步拆分
+- 评估是否需要将 `health/` 模块拆分
+- 考虑添加 `types/` 模块统一管理公共类型
+
+### 2. 文档完善
+- 为每个模块添加详细的文档注释
+- 创建模块依赖关系图
+- 添加架构设计文档
+
+### 3. 测试策略
+- 为每个模块添加单元测试
+- 添加模块间的集成测试
+- 确保重构后所有测试通过
+
+## 总结
+
+通过这次项目结构优化，我们实现了：
+
+1. **文件组织更清晰**：根目录文件从 9 个减少到 2 个
+2. **模块职责明确**：核心功能和通用功能分离
+3. **可维护性提升**：相关功能集中管理，便于维护
+4. **符合最佳实践**：遵循 Rust 项目结构和模块化原则
+5. **保持向后兼容**：通过重新导出保持 API 兼容性
+
+优化后的项目结构更加鲁棒，为项目的长期发展奠定了良好的基础。
