+++ src/core/mod.rs
@@ -1,0 +1,14 @@
+//! 核心模块
+//!
+//! 包含应用程序的核心逻辑和生命周期管理
+
+pub mod app;
+pub mod service;
+pub mod daemon_service;
+pub mod foreground_service;
+
+// 重新导出主要类型
+pub use app::execute_command;
+pub use service::{ServiceManager, ServiceLauncher, ServiceComponents};
+pub use daemon_service::DaemonService;
+pub use foreground_service::ForegroundService;
